Use rollup and babel for ES6 compilation and bundling.
This gets rid of the `lib/` and `test/lib/` folders, since there didn’t seem to be any real use for them. It also switches the `string-selection` dependency to use a real ES6 import, which rollup can understand and use to pull it out of `node_modules/`.

I intentionally left `InputSim` as being referenced globally in the tests because I wanted to test the actual file that was built by the build system to be tested as-is.